scope.c: use mutex_lock_killable to prevent unkillable processes
When doing rmmod quickly after insmod, a race condition causes
the code to wait for the lock forever so it hangs until the next reboot.

This commit changes mutex_lock to mutex_lock_killable so at least
the user can kill the userspace process waiting for the driver.

Even though this is not a permanent proper fix, it makes the issue a lot
less annoying because it removes the need for a full system restart when this happens.